Add polymorphism in basis.
The motivation was to streamline typechecking for head, tail, and if.  This uses
direct substitution, since top-level ascription is mandatory.  Since types are
certain at the top of the AST, and the only mechanism is application, this can't
lead to elaboration getting stuck or confused.

This could be exposed at the top level, but for now I only care about using it
to simplify elaboration in order to root out incorrect LP equations.