Submit each function to the solver separately.
For now, this requires that all functions have the same initial constraints,
because they are solved in no particular order.  If they can be broken down into
strongly-connected components (mutual recursion), then each component could be
solved in order of least dependency.  Because that is only an efficiency
improvement and doesn't impact correctness, it is not a priority.