Continue to use same cost metric to type calls outside of SCC.
This was a mistake based on a misunderstanding of the algorithm description; it
states that the SCC should be analyzed, but it doesn't specify the cost metric.
Obviously, if you use a costly metric for another SCC in the cost-free case of
the current function, that violates the cost-free part.