Split Parser into Reader and Sexp Parser, add Symbol literals.
To add support for `quote` it made more sense to split it instead of glue in
support.  Being homoiconic, there is a first pass that transforms the input into
an s-expression in a very straightforward way, and then a pass over that
s-expression in preorder that assigns program-syntactic categories.  The special
handling of `quote` is accomplished by book-keeping a way to skip over a subtree
during parsing.

The logic of the program-syntactic parsing did not change much, but three things
were gained:

- Symbol literals can now be introduced.
- Now, arbitrary s-expressions could be parsed as a program separately from
  their textual representation, allowing for the possibility of exposing `eval`
  to the user.
- Much better error messages are possible.

The only gotcha in the latter case is that s-expressions are not tagged with
source file location information, so the line/column reporting is bogus.