Fix shift for non-zero inner indices.
Shift could only handle the indices [], [*], [*,*], etc.  More complex indices
like [[*]] would not be able to be shifted, even though they could be paid for,
with [[],[*]] for example.

It is still not completely clear whether [[*]] could be paid for by [[*],[*]] as
I believe shift moves potential across degrees, and this spans 2 degrees.