Project in car/cdr/cons; build environments one position at a time.
Two fixes that had to go together.  The changes in car/cdr/cons were in response
to two things: one, that the shift operator works translates (j, l) <-> (j::l),
which requires projection for car and cdr; two, that shift only pays for j=0,
and everything else has to pay for itself.

The changes building the environment for an application were due to the
realization that it is not a perfect inverse to unpacking the environment inside
a functin; before application, each formal is calculated with respect to the
previous ones without having to consider subsequent ones.