Split apart typechecking from annotation and constraint derivation.
This has the upside of being certain during annotation of a well-typed program
where all the names are in scope.  It also greatly simplifies each, and makes
them both much easier to read.

The one (temporary) downside is that annotation has to recalculate the types and
inject annotations in a few places because it is given a mostly-untyped program.