Double back to using list of paramaters to Arrow instead of pairs.
This undoes parts of 9a9255; that commit was useful at the time, but turns out
to complicate modeling contexts.

Most of the trouble comes down to difficulty pattern matching into a value,
which Ratl doesn't currently, and likely never will, support.  Because PairTy
and PIndex are not separate from Ty and Index, it makes it cumbersome to write
functions that deal with them as contexts---too many special cases and partial
functions for my taste.  If pattern matching were well-supported, I would likely
have to re-revisit this decision.