Move Ty-specific helpers out of Elab to make a Monoid instance for Ty.
Ty cannot form a completely-lawful Monoid because there are forms that cannot be
unified, but the places where the Monoid is useful are restricted, and this can
only happen if the program has a bug.  In places that handle arbitrary types
(e.g., user input), the Monoid cannot be used.