Abandon Monoid instance for a custom typeclass.
The Monoid behavior turned out to be problematic instead of helpful, but since
the code was already half-written, it made sense to rip out the existing solve
function and replace it with a more canonical and complete solver.

It seemed that this belonged more in Elab than Ty, so back it went.

It still blows up on error, but that will be much easier to fix now.