Cover a missed polymorphic case.
`If` was incomplete---checking both directions would sometimes lead to
alpha-equivalent types being rejected as dissimilar.

`App` might still be incomplete in a similar way, but that is harder to fix
without rewriting `elab` to always use fresh Tyvars, instead of `uid`.

The order of unification in `solve` was swapped as a little hack, because
`solve` is supposed to yield a substitution that can applied directly to `t'`,
but for alpha-equivalent substititions the order was backwards.