Workaround for solver failure.
A certain program at max degree 4 was determined to be infeasible, but at degree
3 was feasible.  If a constraint was added that the variable for the fourth
degree was less than 999999999999999.9, it solved it instantly as zero.

The most robust solution is to simply take the feasible solution of the lowest
degree.  The worst thing that can happen is that the answer is asymptotically
over-approximated, which is not the end of the world.