Ground-up rewrite of analysis to multivariate formulation.
Superficially, this was to transform the definition of VarEnv and anno

from: [(LNVar, IxEnv)]     and    a -> m (VarEnv, IxEnv, IxEnv)
to:   ([LNVar], CIxEnv)    and    a -> m (VarEnv, IxEnv)

It was not so trivial, though, because it required a complete reunderstanding of
multivariate analysis.  The key concepts are sequential and parallel traces, and
free variables.

To start, the cost of a non-recursive expression is its constant cost (the exact
costs of its instructions), and the cost induced by any free variables within
that expression.  No other cost is possible.

When composing parallel program traces (as in the case of `if` or `case`
expressions), because branches are mutually exclusive, they are independent.  To
find the projection of each branch, one need only augment the costs of the
branch with the zeroes of the free variables from the other branches, and take
the overall cost as the upper bound of each branch.

Sequences are all dependent, although at first it may not seem to be the case.
Although the exact costs of a given expression may not depend on what proceeds
it, in order to "plumb" the potential of each dimension backwards through the
analysis one must fan out projections into cost-free possible worlds for each
index of interest, before compacting them through sharing.

Sadly, there are still a couple niggling bugs that were inevitably introduced,
so true multivariate analysis is not fully complete yet, and a couple example
programs yield bounds that are too low.