Collapse Cons and Nil into Val, and eliminate List.
Motivated by support for pairs, this had the side benefit of unifying and
streamlining the design of the Val type.  The one major caveat, as noted above
the Elab instance of Val, are restrictions on literals of unit and pair types,
due to syntactic ambiguities with list.  This may be lifted at a later date, if
it's useful.

As another side benefit, the change illuminated a deeper symmetry of
homoiconicity, namely that structural similarity in the AST lies along the right
spine at any point; the left branch is usually shallow, and mere nests to a new
layer of structure.  Specifically, from the root, the right spine contains all
the top level declarations.  The left children of the root's right spine contain
as their right spines the definitions of those.  The right spines of the left
children here each contain the contents of the pieces of the definitions.  The
right spine of the bodies contains the top-level expression, and etc.  So, if
you wanted to get information about the program, this can be obtained through a
pattern traversal.  A traversal with a shallow pattern is exactly what the
parser performs to parse the entire program, but would apply equally to querying
the ascribed types of every definition, for example.