Add ForAll type.
Three reasons:
1) Nano ML uses explicit forall as a part of its "type schemes"
2) Haskell includes forall directly as a type.
3) Introducing scoped type variables is impossible without delimiters.

Currently it ought to be impossible to equate quantified types, because there is
no concrete syntax for ascription, and they are only generalized on insertion
into an environment, and instantiated upon lookup.  This restriction is unlikely
to be lifted in the future, but that will depend on what Haskell does.