Add examples for three different parsers.
The main purpose of these examples is to demonstrate the impact of program size
on analysis performance.  They also serve as a useful demonstration of a more
significant, but still analyzable, program.

These are some cross between LL(1) and recursive descent parsers.  The most
significant weirdness is that in order to derive bounds, `input` *must* be
carried through an entire call stack.  Without this, the epsilon productions
cause `parse` to become infeasible.